sql: Migration to BigInt
Telegram changed from int to long, in which we use BigInt in postgres to please it.

Migration :
ALTER TABLE welcome_restirectlist ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE users ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE chat_members ALTER COLUMN priv_chat_id SET DATA TYPE BigInt;
ALTER TABLE chat_members ALTER COLUMN "user" SET DATA TYPE BigInt;
ALTER TABLE user_report_settings ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE connection ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE connection_history5 ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE userinfo ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE userbio ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE antiflood ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE feds_settings ALTER COLUMN user_id SET DATA TYPE BigInt;
ALTER TABLE warns ALTER COLUMN user_id SET DATA TYPE BigInt;

As well as removing rss sql too, that thing is gone for long.

+++ haruka/modules/sql/connection_sql.py
@@ -18,7 +18,7 @@
 import threading
 from typing import Union
 
-from sqlalchemy import Column, String, Boolean, Integer
+from sqlalchemy import Column, String, Boolean, Integer, BigInteger
 
 from haruka.modules.sql import SESSION, BASE
 
@@ -37,7 +37,7 @@ class ChatAccessConnectionSettings(BASE):
 
 class Connection(BASE):
     __tablename__ = "connection"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id = Column(String(14))
 
     def __init__(self, user_id, chat_id):
@@ -47,7 +47,7 @@ class Connection(BASE):
 
 class ConnectionHistory(BASE):
     __tablename__ = "connection_history5"
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     chat_id1 = Column(String(14))
     chat_id2 = Column(String(14))
     chat_id3 = Column(String(14))
